+++ packages/form/lib/src/extensions.dart
@@ -2,3 +2,4 @@ library extensions;
 
 export 'extensions/form_controller_extension.dart';
 export 'extensions/form_state_field_extension.dart';
+export 'extensions/form_state_value_extension.dart';